ner_extractor/requirements.txt
aiokafka>=0.10.0
asyncpg>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
PyYAML>=6.0.1
jsonschema>=4.21.1
aiohttp>=3.9.0
prometheus-client>=0.20.0
torch==2.0.1
transformers==4.36.0
natasha>=1.6.0
pymorphy2>=0.9.1
